requirements.txt
numpy>=1.24
pandas>=2.0
scipy>=1.10
scikit-learn>=1.3
xgboost>=2.0
matplotlib>=3.7
seaborn>=0.12
yfinance>=0.2.30
stooq @ git+https://github.com/gbournique/stooq.git ; python_version >= "3.10"
pandas-datareader>=0.10
statsmodels>=0.14
requests>=2.31
python-dotenv>=1.0
tqdm>=4.66
pyportfolioopt>=1.5
cvxpy>=1.4
xlrd>=2.0
openpyxl>=3.1
